+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Views
 Extensions
 
 - `Range.init(lower:upper:)` and `ClosedRange.init(lower:upper:)` initializers are added
+- `Collection.containsAnyItem(fromCollection:)` method is added, that indicates if `Collection` contains any item from other `Collection`
 - `UIDevice.safeAreaInsets` no longer returns an `Optional`
 
 ### [4.10.0(83)](https://github.com/VakhoKontridze/VCore/releases/tag/4.10.0) — *2023 06 01*